Rename sourcecode files to not include the word Stream
