+++ OP2Utility.vcxproj
@@ -150,19 +150,19 @@
     <ClCompile Include="src\Archives\HuffLZ.cpp" />
     <ClCompile Include="src\Archives\VolFile.cpp" />
     <ClCompile Include="src\Streams\FileSliceReader.cpp" />
-    <ClCompile Include="src\Streams\FileStreamReader.cpp" />
-    <ClCompile Include="src\Streams\FileStreamWriter.cpp" />
-    <ClCompile Include="src\Streams\MemoryStreamWriter.cpp" />
-    <ClCompile Include="src\Streams\MemoryStreamReader.cpp" />
+    <ClCompile Include="src\Streams\FileReader.cpp" />
+    <ClCompile Include="src\Streams\FileWriter.cpp" />
+    <ClCompile Include="src\Streams\MemoryWriter.cpp" />
+    <ClCompile Include="src\Streams\MemoryReader.cpp" />
     <ClInclude Include="src\Streams\FileSliceReader.h" />
-    <ClInclude Include="src\Streams\FileStreamReader.h" />
-    <ClInclude Include="src\Streams\FileStreamWriter.h" />
-    <ClInclude Include="src\Streams\MemoryStreamReader.h" />
-    <ClInclude Include="src\Streams\MemoryStreamWriter.h" />
-    <ClInclude Include="src\Streams\SeekableStreamReader.h" />
-    <ClInclude Include="src\Streams\SeekableStreamWriter.h" />
-    <ClInclude Include="src\Streams\StreamReader.h" />
-    <ClInclude Include="src\Streams\StreamWriter.h" />
+    <ClInclude Include="src\Streams\FileReader.h" />
+    <ClInclude Include="src\Streams\FileWriter.h" />
+    <ClInclude Include="src\Streams\MemoryReader.h" />
+    <ClInclude Include="src\Streams\MemoryWriter.h" />
+    <ClInclude Include="src\Streams\SeekableReader.h" />
+    <ClInclude Include="src\Streams\SeekableWriter.h" />
+    <ClInclude Include="src\Streams\Reader.h" />
+    <ClInclude Include="src\Streams\Writer.h" />
     <ClInclude Include="src\StringHelper.h" />
     <ClInclude Include="src\XFile.h" />
     <ClCompile Include="src\StringHelper.cpp" />
